packages were renamed: hierarcy -> model list -> util
toString method for classes in model package was made
constructors sequence in model package was fixed
CustomList imports were segregated
.gitignore file was fixed
