+++ .gitignore
@@ -1,2 +1,2 @@
-/.idea
+.idea
 target
