+++ src/main/java/com/model/Blender.java
@@ -1,4 +1,4 @@
-package com.hierarcy;
+package com.model;
 
 import java.time.LocalDate;
 import java.time.Period;
@@ -9,14 +9,14 @@ public class Blender extends Electric{
     private int power;
     private int numberOfMods;
 
+    public Blender() {
+    }
+
     public Blender(int weight, String colour, int price, LocalDate dateOfManufacture, Period guaranteePeriod, int volume, int power, int numberOfMods) {
         super(weight, colour, price, dateOfManufacture, guaranteePeriod);
         this.volume = volume;
         this.power = power;
         this.numberOfMods = numberOfMods;
-    }
-
-    public Blender() {
     }
 
     public int getVolume() {
@@ -55,5 +55,14 @@ public class Blender extends Electric{
     @Override
     public int hashCode() {
         return Objects.hash(super.hashCode(), getVolume(), getPower(), getNumberOfMods());
+    }
+
+    @Override
+    public String toString() {
+        return "Blender{" +
+                "volume=" + volume +
+                ", power=" + power +
+                ", numberOfMods=" + numberOfMods +
+                '}';
     }
 }
