+++ src/test/java/com/util/CustomListTest.java
@@ -1,4 +1,4 @@
-package com.list;
+package com.util;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
